hw11 bug fix

+++ hw11/src/myproject/thread.h
@@ -29,7 +29,8 @@ class Thread {
         }
 
         ~Thread(){
-            CHECK(kill(pid, SIGTERM));
+            if (joinable())
+                CHECK(kill(pid, SIGTERM));
         }
 
         int join(){
